+++ src/components/Inputs/Radio/Radio.tsx
@@ -1,26 +1,36 @@
-import { FormControl, FormControlLabel, FormLabel, Radio as RadioDS, RadioGroup } from "design-system-tracktor";
-import { ChangeEvent, forwardRef, Ref } from "react";
+import { Alert, FormControl, FormControlLabel, FormHelperText, FormLabel, Radio as RadioDS, RadioGroup } from "design-system-tracktor";
+import { ChangeEvent, forwardRef, Ref, useState } from "react";
+import type { ChangeEventField } from "@/features/TreegeForm/type";
 import useInputs from "@/hooks/useInputs";
 import type { TreeNode } from "@/types/TreeNode";
 
 export interface TextFieldProps {
   data: TreeNode;
+  helperText?: string;
   inputRef: Ref<any>;
   required?: boolean;
-  onChange?(event: ChangeEvent<HTMLInputElement>): void;
+  onChange?(dataAttribute: ChangeEventField): void;
 }
 
-const Radio = ({ data, inputRef, required, onChange }: TextFieldProps, ref: Ref<HTMLDivElement>) => {
-  const { getOptionsForDecisionsField } = useInputs();
+const Radio = ({ data, helperText, inputRef, required, onChange }: TextFieldProps, ref: Ref<HTMLDivElement>) => {
+  const { getOptionsForDecisionsField, getMessageByValue } = useInputs();
   const { name, children, attributes } = data;
-  const { label, values } = attributes;
+  const { label, values, type, isLeaf } = attributes;
+  const [message, setMessage] = useState<string | undefined>("");
 
   const options = getOptionsForDecisionsField({ children, values });
+
+  const handleChange = (event: ChangeEvent<HTMLInputElement>, value: string) => {
+    const messageValue = getMessageByValue({ options, value });
+
+    onChange?.({ event, hasMessage: !!messageValue, isLeaf, name, type, value });
+    setMessage(messageValue);
+  };
 
   return (
     <FormControl required={required} ref={ref} fullWidth>
       <FormLabel id={`${name}-label`}>{label}</FormLabel>
-      <RadioGroup aria-labelledby={`${name}-label`} name={name} onChange={onChange} defaultValue="">
+      <RadioGroup aria-labelledby={`${name}-label`} name={name} onChange={handleChange} defaultValue="">
         {options?.map((option, index) => (
           <FormControlLabel
             key={option.key}
@@ -30,6 +40,12 @@ const Radio = ({ data, inputRef, required, onChange }: TextFieldProps, ref: Ref<
           />
         ))}
       </RadioGroup>
+      {helperText && <FormHelperText>{helperText}</FormHelperText>}
+      {message && (
+        <Alert severity="info" variant="standard" sx={{ mt: 1 }}>
+          {message}
+        </Alert>
+      )}
     </FormControl>
   );
 };