+++ src/components/Inputs/Select/Select.tsx
@@ -1,21 +1,40 @@
-import { FormControl, InputLabel, MenuItem, OutlinedInput, Select as SelectDS, SelectChangeEvent } from "design-system-tracktor";
-import { forwardRef, Ref } from "react";
+import {
+  Alert,
+  FormControl,
+  FormHelperText,
+  InputLabel,
+  MenuItem,
+  OutlinedInput,
+  Select as SelectDS,
+  SelectChangeEvent,
+} from "design-system-tracktor";
+import { forwardRef, Ref, useState } from "react";
+import type { ChangeEventField } from "@/features/TreegeForm/type";
 import useInputs from "@/hooks/useInputs";
 import type { TreeNode } from "@/types/TreeNode";
 
 export interface TextFieldProps {
   data: TreeNode;
+  helperText?: string;
   inputRef: Ref<any>;
   required?: boolean;
-  onChange?(event: SelectChangeEvent): void;
+  onChange?(dataAttribute: ChangeEventField): void;
 }
 
-const Select = ({ data, inputRef, required, onChange }: TextFieldProps, ref: Ref<HTMLDivElement>) => {
-  const { getOptionsForDecisionsField } = useInputs();
+const Select = ({ data, helperText, inputRef, required, onChange }: TextFieldProps, ref: Ref<HTMLDivElement>) => {
+  const { getOptionsForDecisionsField, getMessageByValue } = useInputs();
   const { name, children, attributes } = data;
-  const { label, values } = attributes;
-
+  const { label, values, type, isLeaf } = attributes;
+  const [message, setMessage] = useState<string | undefined>("");
   const options = getOptionsForDecisionsField({ children, values });
+
+  const handleChange = (event: SelectChangeEvent) => {
+    const { value } = event.target;
+    const messageValue = getMessageByValue({ options, value });
+
+    onChange?.({ event, hasMessage: !!messageValue, isLeaf, name, type, value });
+    setMessage(messageValue);
+  };
 
   return (
     <FormControl required={required} ref={ref} fullWidth>
@@ -27,7 +46,7 @@ const Select = ({ data, inputRef, required, onChange }: TextFieldProps, ref: Ref
         id={name}
         label={label}
         name={name}
-        onChange={onChange}
+        onChange={handleChange}
         defaultValue=""
         inputRef={inputRef}
         input={<OutlinedInput notched label={label} />}
@@ -38,6 +57,12 @@ const Select = ({ data, inputRef, required, onChange }: TextFieldProps, ref: Ref
           </MenuItem>
         ))}
       </SelectDS>
+      {helperText && <FormHelperText>{helperText}</FormHelperText>}
+      {message && (
+        <Alert severity="info" variant="standard" sx={{ mt: 1 }}>
+          {message}
+        </Alert>
+      )}
     </FormControl>
   );
 };