+++ src/features/TreegeForm/constants/fieldMessageTypes.ts
@@ -1,0 +1,3 @@
+const fieldMessageTypes = ["select", "radio", "switch", "checkbox"];
+
+export default fieldMessageTypes;