+++ src/types/TreeNode.ts
@@ -4,11 +4,14 @@ export interface TreeNodeField {
   isLeaf?: boolean;
   isRoot?: boolean;
   label: string;
+  helperText?: string;
+  messages?: { on: string; off: string };
   required?: boolean;
   step?: string;
   type: string;
   value?: never;
-  values?: { id: string; label: string; value: string }[];
+  values?: { id: string; label: string; value: string; message?: string }[];
+  message?: never;
 }
 
 export interface TreeNodeValues {
@@ -17,11 +20,14 @@ export interface TreeNodeValues {
   isLeaf?: boolean;
   isRoot?: never;
   label: string;
+  helperText?: string;
+  messages?: never;
   required?: never;
   step?: never;
   type?: never;
   value: string;
   values?: never;
+  message?: string;
 }
 
 export type TreeNodeAttributes = TreeNodeField | TreeNodeValues;