+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # [Versions](https://github.com/Tracktor/treege-consumer/releases)
 
-## v1.0.5
-- update dependencies
-- config eslint
+## v1.1.0
+
+- (refactor) Checkbox & switch to boolean field refactor
+- (feat) Add helper message on all fields
+- (feat) Add message to field
+    - Select
+    - Radio
+    - BooleanField
+- (feat) Auto next step for "Select", "Radio" & "Boolean field"